+++ lib/protocol/multipart.rb
@@ -5,6 +5,7 @@
 
 require_relative "multipart/version"
 require_relative "multipart/header"
+require_relative "multipart/headers"
 require_relative "multipart/parser"
 require_relative "multipart/mixed"
 require_relative "multipart/part"