+++ releases.md
@@ -2,6 +2,6 @@
 
 ## Unreleased
 
-  - Add `Protocol::Multipart::Header` for parsing parameterized header values.
+  - Add `Protocol::Multipart::Header` for parsing parameterized header values and `Protocol::Multipart::Headers` for storing multipart header fields.
 
 ## v0.1.0